requirements.txt
aiofiles==0.8.0
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coverage==6.4.1
cryptography==37.0.2
ecdsa==0.17.0
fastapi==0.78.0
fastapi-sessions==0.3.2
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
mailchimp-transactional==1.0.47
MouseInfo==0.1.3
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psutil==5.9.1
py==1.11.0
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==1.9.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytest==7.1.2
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.5
pytweening==1.0.4
requests==2.28.0
rsa==4.8
rubicon-objc==0.4.2
sendgrid==6.9.7
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starkbank-ecdsa==2.0.3
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
